Import site from www.rust-bitcoin.org
Import the source code for the website (and cookbook) from the master
branch of `github.com/rust-bitcoin/www.rust-bitcoin.org` at tip:

    commit 2f3bbbf3c43afac377aad9163b6a1b0bb871b18e

This includes `.github/`, `.gitmodules`, an `.gitignore` but _not_
`.git`.

The repo is not currently configured with write permissions to the
`gh-pages` branch so pushing this patch to master will not make any
changes to the branch currently where the site is currently being served
from.